+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Fixed
+
+- Error when `VersionUpdate` is done when a new model has no version yet in the model registry
+
 ## v2.11.1 - 2023-02-03
 
 ### Fixed